Moved all info to Keyv and more
As the title suggests, I moved all the info stored in the .json files to Keyv. I also added a data command that only I can use, as well as make some code a lot more legible. I'm still working on dynamic errors, that will probably be in the next commit along with the tourney command (#6).