I'm still working on this project (committed from and resolved #13)
Keyv screwed a ton of stuff up so i'm gonna use .jsons and if that doesnt work i'll use sqlite or something. right now i'm working on #6 (tourney command). next on my list from highest priority to lowest is #9 (HSE Achievements) #8 (wiki) and #12 (copypasta). 

footnote: THE CODE HERE IS NOT FULLY FUNCTIONING! IT HAS A FEW FATAL FLAWS AND IT HAS UNFINISHED FUNCTIONS.