v1.4.2
I know it's been said a lot, but the tourney command is almost done. I just have to get the bracket mechanics working. Also, /poll and /copypasta are ready! The wiki and achievements commands will be next. One more thing: BruhBot will now enforce the banned words as defined in #rules-and-info. This commit resolves issues #12 and #14.