merge un-needed branch and delete (#15)
* v1.4.0

Fixed some bugs, added a node module or 2, and got the base of `/tourney` set up

* v1.4.1 

Updated `/tourney`, it's closer to being finished! Also updated `/role` so that it supports interactive mode. Music will be added soon, so keep an eye out for that!

* v1.4.2

I know it's been said a lot, but the tourney command is almost done. I just have to get the bracket mechanics working. Also, /poll and /copypasta are ready! The wiki and achievements commands will be next. This commit resolves issues #12 and #14.